requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# MLP_final_project\src\api.py: 1
fastapi==0.88.0
# MLP_final_project\data_prepocessing.ipynb: 4
imbalanced-learn==0.9.1
# MLP_final_project\eda.ipynb: 11
ipython==7.24.1
# MLP_final_project\src\util.py: 2
joblib==1.1.0
# MLP_final_project\eda.ipynb: 2
# MLP_final_project\modelling.ipynb: 3
matplotlib==3.4.2
# MLP_final_project\data_prepocessing.ipynb: 3
# MLP_final_project\playground.ipynb: 3
# MLP_final_project\src\api.py: 5
numpy==1.20.3
# MLP_final_project\data_preparation.ipynb: 1
# MLP_final_project\data_prepocessing.ipynb: 2
# MLP_final_project\eda.ipynb: 4
# MLP_final_project\modelling.ipynb: 2
# MLP_final_project\playground.ipynb: 2
# MLP_final_project\src\api.py: 4
# MLP_final_project\src\data_preprocessing.py: 1
pandas==1.4.2
# MLP_final_project\src\streamlit.py: 3
Pillow==8.2.0
# MLP_final_project\src\api.py: 2
pydantic==1.10.2
# MLP_final_project\src\util.py: 1
PyYAML==6.0
# MLP_final_project\src\streamlit.py: 2
requests==2.28.1
# MLP_final_project\data_preparation.ipynb: 3
# MLP_final_project\data_prepocessing.ipynb: 5
# MLP_final_project\modelling.ipynb: 7,8,10,11,12,13,14,15,16,17,18
# MLP_final_project\playground.ipynb: 4
scikit-learn==1.1.3
# MLP_final_project\eda.ipynb: 1
seaborn==0.12.1
# MLP_final_project\src\streamlit.py: 1
streamlit==1.16.0
# MLP_final_project\src\api.py: 3
uvicorn==0.20.0
# MLP_final_project\modelling.ipynb: 19
xgboost==1.7.1
